# cometmind/openapi.yaml
openapi: 3.1.0
info:
  title: CometMind Local Serve API
  version: 0.1.0
  description: |
    Minimal local HTTP + SSE surface for CometMind Phase 0.

    v1 is localhost-only, has no app-layer auth, keeps `/api/v1` path versioning,
    and exposes only the first serve slice endpoints.
servers:
  - url: http://127.0.0.1:7700
    description: Default local bind address
...
paths:
...
  /api/v1/mcp/servers:
    get:
      tags: [MCP]
      summary: List configured MCP servers and runtime status
      operationId: listMcpServers
      responses:
        '200':
          description: MCP server statuses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMcpServersResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      operationId: listJobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, ongoing, done]
        - name: ready_only
          in: query
          schema:
            type: boolean
        - name: include_deleted
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Job list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListJobsResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Jobs]
      summary: Create a job
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get a job
      operationId: getJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResource'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Jobs]
      summary: Update a todo job
      operationId: updateJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobRequest'
      responses:
        '200':
          description: Updated job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Jobs]
      summary: Soft-delete a job
      operationId: deleteJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: bad_request
              message: invalid request
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: session_not_found
              message: session was not found
...
    Conflict:
      description: State conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: conflict
              message: resource state conflict
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: internal_error
              message: unexpected server error
  schemas:
...
    McpServerStatus:
      type: object
      additionalProperties: false
      required: [id, name, enabled, transport, status, tool_count]
      properties:
        id:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        transport:
          type: string
        status:
          type: string
          enum: [disabled, connected, error, disconnected]
        tool_count:
          type: integer
        last_error:
          type: string
        oauth_connected:
          type: boolean
    ListMcpServersResponse:
      type: object
      additionalProperties: false
      required: [servers]
      properties:
        servers:
          type: array
          items:
            $ref: '#/components/schemas/McpServerStatus'
...
    JobResource:
      type: object
      additionalProperties: false
      required:
        - id
        - description
        - definition_of_done
        - progress
        - status
        - priority
        - created_by
        - created_at
        - updated_at
      properties:
        id:
          type: string
        description:
          type: string
        definition_of_done:
          type: string
        progress:
          type: string
        status:
          type: string
          enum: [todo, ongoing, done]
        priority:
          type: integer
        scheduled_at:
          type: integer
          format: int64
        due_at:
          type: integer
          format: int64
        workspace_path:
          type: string
        assigned_session_id:
          type: string
        lease_expires_at:
          type: integer
          format: int64
        created_by:
          type: string
        source_session_id:
          type: string
        source_platform:
          type: string
        source_channel_id:
          type: string
        deleted_at:
          type: integer
          format: int64
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
...
    ListJobsResponse:
      type: object
      additionalProperties: false
      required: [jobs]
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobResource'
...
    CreateJobRequest:
      type: object
      additionalProperties: false
      required: [description]
      properties:
        description:
          type: string
        definition_of_done:
          type: string
        priority:
          type: integer
        scheduled_at:
          type: integer
          format: int64
        due_at:
          type: integer
          format: int64
        workspace_path:
          type: string
        created_by:
          type: string
        source_session_id:
          type: string
        source_platform:
          type: string
        source_channel_id:
          type: string
    UpdateJobRequest:
      type: object
      additionalProperties: false
      required: [description, definition_of_done, priority]
      properties:
        description:
          type: string
        definition_of_done:
          type: string
        priority:
          type: integer
        scheduled_at:
          type: integer
          format: int64
        due_at:
          type: integer
          format: int64
        workspace_path:
          type: string
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string